+++ x-pack/test/api_integration/apis/infra/feature_controls.ts
@@ -6,7 +6,6 @@
 
 import expect from '@kbn/expect';
 import gql from 'graphql-tag';
-import { SecurityService, SpacesService } from '../../../common/services';
 import { FtrProviderContext } from '../../ftr_provider_context';
 
 const introspectionQuery = gql`
@@ -21,8 +20,8 @@ const introspectionQuery = gql`
 
 export default function({ getService }: FtrProviderContext) {
   const supertest = getService('supertestWithoutAuth');
-  const security: SecurityService = getService('security');
-  const spaces: SpacesService = getService('spaces');
+  const security = getService('security');
+  const spaces = getService('spaces');
   const clientFactory = getService('infraOpsGraphQLClientFactory');
 
   const expectGraphQL404 = (result: any) => {
