+++ x-pack/test/api_integration/apis/security/change_password.ts
@@ -6,11 +6,10 @@
 
 import { Cookie, cookie } from 'request';
 import { FtrProviderContext } from '../../ftr_provider_context';
-import { SecurityService } from '../../../common/services/security';
 
 export default function({ getService }: FtrProviderContext) {
   const supertest = getService('supertestWithoutAuth');
-  const security: SecurityService = getService('security');
+  const security = getService('security');
 
   const mockUserName = 'test-user';
   const mockUserPassword = 'test-password';
