+++ x-pack/test/api_integration/apis/siem/feature_controls.ts
@@ -6,7 +6,6 @@
 
 import expect from '@kbn/expect';
 import gql from 'graphql-tag';
-import { SecurityService, SpacesService } from '../../../common/services';
 import { FtrProviderContext } from '../../ftr_provider_context';
 
 const introspectionQuery = gql`
@@ -22,8 +21,8 @@ const introspectionQuery = gql`
 export default function({ getService }: FtrProviderContext) {
   const config = getService('config');
   const supertest = getService('supertestWithoutAuth');
-  const security: SecurityService = getService('security');
-  const spaces: SpacesService = getService('spaces');
+  const security = getService('security');
+  const spaces = getService('spaces');
   const clientFactory = getService('siemGraphQLClientFactory');
 
   const expectGraphQL404 = (result: any) => {
