+++ x-pack/test/api_integration/services/index.ts
@@ -5,9 +5,7 @@
  */
 
 import { services as kibanaApiIntegrationServices } from '../../../../test/api_integration/services';
-import { services as kibanaCommonServices } from '../../../../test/common/services';
-
-import { SecurityServiceProvider, SpacesServiceProvider } from '../../common/services';
+import { services as commonServices } from '../../common/services';
 
 // @ts-ignore not ts yet
 import { LegacyEsProvider } from './legacy_es';
@@ -25,7 +23,7 @@ import { SiemGraphQLClientProvider, SiemGraphQLClientFactoryProvider } from './s
 import { InfraOpsSourceConfigurationProvider } from './infraops_source_configuration';
 
 export const services = {
-  ...kibanaCommonServices,
+  ...commonServices,
 
   esSupertest: kibanaApiIntegrationServices.esSupertest,
   supertest: kibanaApiIntegrationServices.supertest,
@@ -35,10 +33,8 @@ export const services = {
   infraOpsGraphQLClient: InfraOpsGraphQLClientProvider,
   infraOpsGraphQLClientFactory: InfraOpsGraphQLClientFactoryProvider,
   infraOpsSourceConfiguration: InfraOpsSourceConfigurationProvider,
-  security: SecurityServiceProvider,
   siemGraphQLClient: SiemGraphQLClientProvider,
   siemGraphQLClientFactory: SiemGraphQLClientFactoryProvider,
-  spaces: SpacesServiceProvider,
   supertestWithoutAuth: SupertestWithoutAuthProvider,
   usageAPI: UsageAPIProvider,
 };
