+++ x-pack/test/common/services/index.ts
@@ -4,5 +4,14 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export { SecurityServiceProvider, SecurityService } from './security';
-export { SpacesServiceProvider, SpacesService } from './spaces';
+import { services as kibanaCommonServices } from '../../../../test/common/services';
+
+import { SecurityServiceProvider } from './security';
+import { SpacesServiceProvider } from './spaces';
+
+export const services = {
+  ...kibanaCommonServices,
+
+  security: SecurityServiceProvider,
+  spaces: SpacesServiceProvider,
+};
