+++ x-pack/test/functional/apps/spaces/copy_saved_objects.ts
@@ -4,7 +4,6 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import expect from '@kbn/expect';
-import { SpacesService } from '../../../common/services';
 import { FtrProviderContext } from '../../ftr_provider_context';
 
 export default function spaceSelectorFunctonalTests({
@@ -12,7 +11,7 @@ export default function spaceSelectorFunctonalTests({
   getPageObjects,
 }: FtrProviderContext) {
   const esArchiver = getService('esArchiver');
-  const spaces: SpacesService = getService('spaces');
+  const spaces = getService('spaces');
   const testSubjects = getService('testSubjects');
   const PageObjects = getPageObjects(['security', 'settings', 'copySavedObjectsToSpace']);
