+++ x-pack/test/functional/services/index.ts
@@ -6,6 +6,7 @@
 
 import { services as kibanaFunctionalServices } from '../../../../test/functional/services';
 import { services as kibanaApiIntegrationServices } from '../../../../test/api_integration/services';
+import { services as commonServices } from '../../common/services';
 
 import {
   MonitoringNoDataProvider,
@@ -49,13 +50,13 @@ import { InfraLogStreamProvider } from './infra_log_stream';
 import { MachineLearningProvider } from './ml';
 import { TransformProvider } from './transform';
 
-import { SecurityServiceProvider, SpacesServiceProvider } from '../../common/services';
-
 // define the name and providers for services that should be
 // available to your tests. If you don't specify anything here
 // only the built-in services will be available
 export const services = {
   ...kibanaFunctionalServices,
+  ...commonServices,
+
   esSupertest: kibanaApiIntegrationServices.esSupertest,
   monitoringNoData: MonitoringNoDataProvider,
   monitoringClusterList: MonitoringClusterListProvider,
@@ -83,8 +84,6 @@ export const services = {
   random: RandomProvider,
   aceEditor: AceEditorProvider,
   grokDebugger: GrokDebuggerProvider,
-  security: SecurityServiceProvider,
-  spaces: SpacesServiceProvider,
   userMenu: UserMenuProvider,
   uptime: UptimeProvider,
   infraSourceConfigurationForm: InfraSourceConfigurationFormProvider,
