+++ x-pack/test/saml_api_integration/services.ts
@@ -4,9 +4,11 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { services as commonServices } from '../common/services';
 import { services as apiIntegrationServices } from '../api_integration/services';
 
 export const services = {
+  ...commonServices,
   randomness: apiIntegrationServices.randomness,
   legacyEs: apiIntegrationServices.legacyEs,
   supertestWithoutAuth: apiIntegrationServices.supertestWithoutAuth,
