+++ x-pack/test/saved_object_api_integration/common/services/index.ts
@@ -7,11 +7,13 @@
 // @ts-ignore not ts yet
 import { LegacyEsProvider } from './legacy_es';
 
+import { services as commonServices } from '../../../common/services';
 import { services as apiIntegrationServices } from '../../../api_integration/services';
 import { services as kibanaApiIntegrationServices } from '../../../../../test/api_integration/services';
 import { services as kibanaFunctionalServices } from '../../../../../test/functional/services';
 
 export const services = {
+  ...commonServices,
   legacyEs: LegacyEsProvider,
   esSupertestWithoutAuth: apiIntegrationServices.esSupertestWithoutAuth,
   supertest: kibanaApiIntegrationServices.supertest,
