+++ x-pack/test/ui_capabilities/security_and_spaces/tests/index.ts
@@ -4,16 +4,14 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { SpacesService } from '../../../common/services';
-import { SecurityService } from '../../../common/services';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { FeaturesService } from '../../common/services';
 import { isCustomRoleSpecification } from '../../common/types';
 import { Spaces, Users } from '../scenarios';
 
 export default function uiCapabilitiesTests({ loadTestFile, getService }: FtrProviderContext) {
-  const securityService: SecurityService = getService('security');
-  const spacesService: SpacesService = getService('spaces');
+  const securityService = getService('security');
+  const spacesService = getService('spaces');
   const featuresService: FeaturesService = getService('features');
 
   describe('ui capabilities', function() {
