+++ x-pack/test/ui_capabilities/security_only/tests/index.ts
@@ -4,13 +4,12 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { SecurityService } from '../../../common/services';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { isCustomRoleSpecification } from '../../common/types';
 import { UserScenarios } from '../scenarios';
 
 export default function uiCapabilitesTests({ loadTestFile, getService }: FtrProviderContext) {
-  const securityService: SecurityService = getService('security');
+  const securityService = getService('security');
 
   describe('ui capabilities', function() {
     this.tags('ciGroup9');
