+++ x-pack/test/ui_capabilities/spaces_only/tests/index.ts
@@ -4,13 +4,12 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { SpacesService } from '../../../common/services';
 import { FtrProviderContext } from '../../common/ftr_provider_context';
 import { FeaturesService } from '../../common/services';
 import { SpaceScenarios } from '../scenarios';
 
 export default function uiCapabilitesTests({ loadTestFile, getService }: FtrProviderContext) {
-  const spacesService: SpacesService = getService('spaces');
+  const spacesService = getService('spaces');
   const featuresService: FeaturesService = getService('features');
 
   describe('ui capabilities', function() {
